requirements.txt
asgiref==3.4.1
bleach==4.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
django-markdownify==0.9.0
django-request-logging==0.7.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
packaging==21.0
psycopg2==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
